Feat: sign out and header navigation
