+++ packages/client/src/App.js
@@ -2,12 +2,12 @@ import React, { useEffect } from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { useDispatch } from "react-redux";
 
-import { PUBLIC, PRIVATE } from "./constants/routes";
+import { PUBLIC } from "./constants/routes";
 import Home from "./pages/Public/Home";
 import SignUp from "./pages/Public/SignUp";
 import SignIn from "./pages/Public/SignIn";
-import Account from "./pages/Private/Account";
-import ResetPassword from "./pages/Public/ResetPassword/ResetPassword";
+import Account from "./pages/Public/Account";
+// import ResetPassword from "./pages/Public/ResetPassword/ResetPassword";
 import { onAuthStateChanged } from "./services/auth";
 import { logIn } from "./redux/user/actions";
 import { signInUserData } from "./api/account-api";
@@ -42,7 +42,7 @@ function App() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route path={PRIVATE.USER_ACCOUNT}>
+        <Route path={PUBLIC.USER_ACCOUNT}>
           <Account />
         </Route>
         <Route path={PUBLIC.SIGN_UP}>
@@ -51,9 +51,9 @@ function App() {
         <Route path={PUBLIC.SIGN_IN}>
           <SignIn />
         </Route>
-        <Route path={PUBLIC.RESET_PASSWORD}>
+        {/* <Route path={PUBLIC.RESET_PASSWORD}>
           <ResetPassword />
-        </Route>
+        </Route> */}
         <Route path={PUBLIC.HOME}>
           <Home />
         </Route>
