--- packages/client/src/pages/Private/Account/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Account";
