+++ packages/client/src/pages/Public/Account/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Account";
